Resizing

+++ Ivory/src/OrthoCameraController.cpp
@@ -35,8 +35,9 @@ namespace Ivory {
 		return false;
 	}
 	bool OrthographicCameraController::on_window_resize(WindowResizeEvent& e) {
-		m_aspect_ratio = e.get_width() / e.get_height();
+		m_aspect_ratio = (float)e.get_width() / (float)e.get_height();
 		m_camera.set_projection(-m_aspect_ratio * m_zoom_level, m_aspect_ratio * m_zoom_level, -m_zoom_level, m_zoom_level);
+
 		return false;
 	}
 }
